requirements.txt
numpy==1.22.3
Pillow==9.1.0
PyOpenGL==3.1.6
PyQt5==5.15.6
ephem~=4.1.3
googletrans~=3.1.0a0
requests~=2.27.1
